Added Level Collisions
Added collisions for the entire playable level.

Developer note: the collisions on stairs are impossible to work with. The player gets stuck and there is nothing I can do to fix it other than to suggest removing the stairs completely. As bad as it sounds, it would at least fix the mounting issue on ever god damn step